# src/docs/openapi.yaml
openapi: 3.1.0
info:
  title: FisioFlow API
  description: |
    API da FisioFlow — Cloudflare Worker (Hono) + Neon Postgres via Hyperdrive.
    Autenticação: JWT do Neon Auth no header `Authorization: Bearer <token>` (validado via JWKS).
    Os endpoints sob `/api/health/*` são públicos. Paths derivados de `apps/api/src/routes/*`.
  version: 3.0.0
servers:
  - url: https://fisioflow-api.rafalegollas.workers.dev
    description: Produção (Cloudflare Worker)
  - url: http://localhost:8787
    description: Desenvolvimento local (pnpm workers:dev)
security:
  - bearerAuth: []
...
paths:
...
  /api/nfse:
    get:
      tags: [NFS-e]
      summary: Lista NFS-e
      responses:
        "200": { description: Lista de NFS-e }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/nfse/config:
    get:
      tags: [NFS-e]
      summary: Configuração NFS-e
      responses:
        "200": { description: Configuração }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/nfse/test-connection:
    get:
      tags: [NFS-e]
      summary: Testa conexão mTLS com a prefeitura (SP)
      description: Requer auth. Retorna 422 se o certificado não estiver configurado e 502 se o serviço externo falhar.
      responses:
        "200": { description: Conexão OK }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { description: Certificado não configurado }
        "502": { description: Falha no serviço externo }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT do Neon Auth.
...
  responses:
    Unauthorized:
      description: Token ausente ou inválido
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }